Eu te amo muito Carlinhos!! Eu te amo Carlinhos Gustavo Cecyn Lundgren!! Eu te amo Carlos Gustavo Cecyn Lundgren!!
Obrigado Mummi!! Feliz 61 anos!!

Obrigado Carol, Lúcio e Lucinho!!

comments from cvcl repos commit at curitiba remotely from here Santos at this veri machine and also veri affected in portuenglish, i.e. happy!! Because you loved me!! Both…
"Eu te amo Carlinhos!!
Você é o meu Deus, Carlos Gustavo Cecyn Lundgren 1984-Nov-15!!

Obrigado Mummi!! Feliz 61 anos!!

Obrigado Carol, Lúcio e Lucinho!!

Obrigado Rafael Cecyn e Gabi!!

Obrigado Simon e Iara!!

Obrigado K. seu luiz!
"

porting to Mac at OS X Mountain Lion 10.8.2 as VMware guest with 8 cores 3GB RAM at Carolina Kiyoko Dell at ckribeiro LAN at Avenida Bernardino de Campos Santos SP, Brazil

at machine Dell at network NetVirtua 10Mbps at Santos, SP - Brazil, at laptop Dell XPS L502X processor i7 2670QM 8GB RAM 1 TB Disk 1.5TB Backup Samsung Disk at Santos SP, Brazil

-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Léia Maheli de Souza Tsumanuma!! Mummi
  Tack Simon Petrus Lundgren!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabrielli Cecyn Lundgren!! Gabi
  Tack Rafael Cecyn Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
}

Obrigado Cits
{
   Obrigado Rejeane!!
   Obrigado Monique!!
   Obrigado Gilmar Tuchinski!!
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Buhrer!!
}

Obrigado Brasil
{
   Obrigado guardas!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado moças e mulheres...
   Obrigado moços e homens...
   Obrigado Keila Mª da Silva!! Babysitter do Lucinho
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Dilma Souza!!
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}

Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt & Edras Pacola!! 
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you Anderson Silva!!
   Thank you Pietro!!
   Thank you Thiago Ferreira!! - de 1996 do grupo do Cleiton, Anderson Silva e Pietro
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
T Obrigado Amigo e Amiga, você é legal!!
1 Obrigado Amigo e Amiga, você é bom!!
2 Obrigado Amigo e Amiga, eu gosto de você!!
n Obrigado Amigo e Amiga, eu gosto muito de você!!
n Obrigado Amigo e Amiga, você é demais!!
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a desobrigado.
+∞

+++ appseed/openssl/crypto/ec/ec_asn1.c
@@ -520,8 +520,8 @@ static int ec_asn1_group2curve(const EC_GROUP *group, X9_62_CURVE *curve)
 		}
 	
 	/* set a and b */
-	if (!M_ASN1_OCTET_STRING_set(curve->a, a_buf, len_1) ||
-	    !M_ASN1_OCTET_STRING_set(curve->b, b_buf, len_2))
+	if (!M_ASN1_OCTET_STRING_set(curve->a, a_buf, (int) len_1) ||
+	    !M_ASN1_OCTET_STRING_set(curve->b, b_buf, (int) len_2))
 		{
 		ECerr(EC_F_EC_ASN1_GROUP2CURVE, ERR_R_ASN1_LIB);
 		goto err;
@@ -643,7 +643,7 @@ static ECPARAMETERS *ec_asn1_group2parameters(const EC_GROUP *group,
 		ECerr(EC_F_EC_ASN1_GROUP2PARAMETERS, ERR_R_MALLOC_FAILURE);
 		goto err;
 		}
-	if (!ASN1_OCTET_STRING_set(ret->base, buffer, len))
+	if (!ASN1_OCTET_STRING_set(ret->base, buffer, (int) len))
 		{
 		ECerr(EC_F_EC_ASN1_GROUP2PARAMETERS, ERR_R_ASN1_LIB);
 		goto err;
@@ -1159,7 +1159,7 @@ EC_KEY *d2i_ECPrivateKey(EC_KEY **a, const unsigned char **in, long len)
 		goto err;
 		}
 
-	ret->version = priv_key->version;
+	ret->version = (int) priv_key->version;
 
 	if (priv_key->privateKey)
 		{
@@ -1259,7 +1259,7 @@ int	i2d_ECPrivateKey(EC_KEY *a, unsigned char **out)
 		goto err;
 		}
 
-	if (!M_ASN1_OCTET_STRING_set(priv_key->privateKey, buffer, buf_len))
+	if (!M_ASN1_OCTET_STRING_set(priv_key->privateKey, buffer, (int) buf_len))
 		{
 		ECerr(EC_F_I2D_ECPRIVATEKEY, ERR_R_ASN1_LIB);
 		goto err;
@@ -1310,7 +1310,7 @@ int	i2d_ECPrivateKey(EC_KEY *a, unsigned char **out)
 		priv_key->publicKey->flags &= ~(ASN1_STRING_FLAG_BITS_LEFT|0x07);
 		priv_key->publicKey->flags |= ASN1_STRING_FLAG_BITS_LEFT;
 		if (!M_ASN1_BIT_STRING_set(priv_key->publicKey, buffer, 
-				buf_len))
+				(int) buf_len))
 			{
 			ECerr(EC_F_I2D_ECPRIVATEKEY, ERR_R_ASN1_LIB);
 			goto err;
@@ -1418,7 +1418,7 @@ int i2o_ECPublicKey(EC_KEY *a, unsigned char **out)
 
 	if (out == NULL || buf_len == 0)
 	/* out == NULL => just return the length of the octet string */
-		return buf_len;
+		return (int) buf_len;
 
 	if (*out == NULL)
 		{
@@ -1439,5 +1439,5 @@ int i2o_ECPublicKey(EC_KEY *a, unsigned char **out)
 		}
 	if (!new_buffer)
 		*out += buf_len;
-	return buf_len;
+	return (int) buf_len;
 	}
